Create Blue_GPT_Wrapper.py
my first iteration of this. It works well to create the first iteration of Blue, with obvious limitations, no input sanitation, and very little protections. However, it's my first major Python project and makes use of the OpenAI API. 
I love the idea of a command like assistant and removing the friction from using this kind of technology.

This should be considered as an option for personal use as far as OpenAI API wrappers go.